+++ examples/imagepull.rs
@@ -1,13 +1,13 @@
 extern crate shiplift;
 
-use shiplift::Docker;
+use shiplift::{Docker, PullOptions};
 use std::env;
 
 fn main() {
     let docker = Docker::new();
     if let Some(img) = env::args().nth(1) {
         let image = docker.images()
-            .create(&img[..])
+            .pull(&PullOptions::builder().image(img).build())
             .unwrap();
         for output in image {
             println!("{:?}", output);